% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kissr.R
\docType{package}
\name{kissr}
\alias{kissr}
\alias{kissr-package}
\title{kissr: A package for loading report data from Kiss Metrics}
\description{
The kissr package provides two important functions:
read, and list.
}

